requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
reportlab==4.0.4
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9